added more admin functionality + design from another branch

+++ ForumApp/ForumApp/Controllers/UsersController.cs
@@ -9,7 +9,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace ForumApp.Controllers
 {
-    [Authorize(Roles = "Admin")]
+    [Authorize]
     public class UsersController : Controller
     {
         private readonly ApplicationDbContext db;
@@ -26,6 +26,7 @@ namespace ForumApp.Controllers
             _roleManager = roleManager;
         }
 
+        [Authorize(Roles = "Admin")]
         public IActionResult Index()
         {
             var users = from user in db.Users
@@ -36,57 +37,85 @@ namespace ForumApp.Controllers
             return View();
         }
 
+        [Authorize(Roles = "User,Editor,Admin")]
         public async Task<ActionResult> Edit(string id)
         {
             ApplicationUser user = db.Users.Find(id);
-
+            SetAccessRights();
             user.AllRoles = GetAllRoles();
 
-            var roleNames = await _userManager.GetRolesAsync(user);         // lista nume de roluri
+            var roleNames = await _userManager.GetRolesAsync(user);     // listam numele rolurilor
 
+            // cautam id ul rolului in baza de date
             var currentUserRole = _roleManager.Roles
                                               .Where(r => roleNames.Contains(r.Name))
                                               .Select(r => r.Id)
-                                              .First();                     // selectam un singur rol
+                                              .First();     // k selectam un singur rol
+
             ViewBag.UserRole = currentUserRole;
 
-            return View(user);
+            // verificam daca userul care acceseaza viewul editului este admin
+            // sau este userul insusi
+            if(user.Id == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                return View(user);
+            }
+            else
+            {
+                TempData["message"] = "Access denied!";
+                return Redirect("/Sections/Index");
+            }
+
+        }
+
+        private void SetAccessRights()
+        {
+            ViewBag.UserCurent = _userManager.GetUserId(User);
         }
 
         [HttpPost]
-        public async Task<ActionResult> Edit(string id, ApplicationUser newData, [FromForm] string newRole)
+        [Authorize(Roles = "User,Editor,Admin")]
+        public async Task<ActionResult> Edit(string id, ApplicationUser newUser, [FromForm] string newRole)
         {
             ApplicationUser user = db.Users.Find(id);
 
             user.AllRoles = GetAllRoles();
-
-
-            if (ModelState.IsValid)
+            if(ModelState.IsValid)
             {
-                user.UserName = newData.UserName;
-                user.Email = newData.Email;
-                user.FirstName = newData.FirstName;
-                user.LastName = newData.LastName;
-                user.PhoneNumber = newData.PhoneNumber;
 
-
-                // Cautam toate rolurile din baza de date
-                var roles = db.Roles.ToList();
-
-                foreach (var role in roles)
+                // verificam sa nu faca cineva pe hackerul
+                if (user.Id == _userManager.GetUserId(User) || User.IsInRole("Admin"))
                 {
-                    // Scoatem userul din rolurile anterioare
-                    await _userManager.RemoveFromRoleAsync(user, role.Name);
+                    user.UserName = newUser.UserName;
+                    user.Email = newUser.Email;
+                    user.FirstName = newUser.FirstName;
+                    user.LastName = newUser.LastName;
+                    user.PhoneNumber = newUser.PhoneNumber;
+
+                    // cautam rolurile in baza de date
+                    var roles = db.Roles.ToList();
+                    foreach (var role in roles)
+                    {
+                        // scoatem userul din rolul precedent
+                        await _userManager.RemoveFromRoleAsync(user, role.Name);
+                    }
+                    // il unim cu rolul selectat
+                    var roleName = await _roleManager.FindByIdAsync(newRole);
+                    await _userManager.AddToRoleAsync(user, roleName.ToString());
+
+                    db.SaveChanges();
                 }
-
-                // Adaugam noul rol selectat
-                var roleName = await _roleManager.FindByIdAsync(newRole);
-                await _userManager.AddToRoleAsync(user, roleName.ToString());
-
-                db.SaveChanges();
-
             }
-            return RedirectToAction("Index");
+            if(User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                TempData["message"] = "Changes applied successfully!";
+                return Redirect("/Sections/Index");
+            }
+            
         }
 
         public async Task<ActionResult> Show(string Id)
@@ -94,12 +123,15 @@ namespace ForumApp.Controllers
             ApplicationUser user = db.Users.Find(Id);
             var roles = await _userManager.GetRolesAsync(user);
 
-            ViewBag.Roles = roles;
+            SetAccessRights();
 
+            ViewBag.Roles = roles;
             return View(user);
+            
         }
 
         [HttpPost]
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Delete(string Id)
         {
             var user = db.Users
@@ -109,46 +141,58 @@ namespace ForumApp.Controllers
                             .Where(u => u.Id == Id)
                             .First();
 
-            // stergem forumurile create de user
-            if (user.Forums.Count > 0)
+            // daca e user simplu verific sa fie idul lui
+            // daca e admin verific sa nu-si dea singur delete
+            if ((user.Id == _userManager.GetUserId(User) && User.IsInRole("Admin") == false) || (user.Id != _userManager.GetUserId(User) && User.IsInRole("Admin"))) 
             {
-                
-                foreach (var forum in user.Forums)
+                // stergem forumurile create de user
+                if (user.Forums.Count > 0)
                 {
-                    //Subforum sf = db.Subforums.Find(u => u.Id == forum.Id);
 
-                    db.Forums.Remove(forum);
+                    foreach (var forum in user.Forums)
+                    {
+                        //Subforum sf = db.Subforums.Find(u => u.Id == forum.Id);
+
+                        db.Forums.Remove(forum);
+                    }
                 }
+
+                if (user.Subforums.Count > 0)
+                {
+                    foreach (var subforum in user.Subforums)
+                    {
+                        // decrementam numarul de subforumuri
+                        Forum f = db.Forums.Find(subforum.ForumId);
+                        f.CountOfSubforums--;
+                        db.Subforums.Remove(subforum);
+                    }
+                }
+
+                if (user.Posts.Count > 0)
+                {
+                    foreach (var post in user.Posts)
+                    {
+                        // decrementam nr de mesaje din forum si subforum
+                        Subforum sf = db.Subforums.Find(post.SubforumId);
+                        Forum f = db.Forums.Find(sf.ForumId);
+                        sf.MsgCount--;
+                        f.MsgCount--;
+                        db.Posts.Remove(post);
+                    }
+                }
+
+                db.ApplicationUsers.Remove(user);
+                db.SaveChanges();
             }
 
-            if (user.Subforums.Count > 0)
+            if (User.IsInRole("Admin"))
             {
-                foreach (var subforum in user.Subforums)
-                {
-                    // decrementam numarul de subforumuri
-                    Forum f = db.Forums.Find(subforum.ForumId);
-                    f.CountOfSubforums--;
-                    db.Subforums.Remove(subforum);
-                }
+                return RedirectToAction("Index");
             }
-
-            if (user.Posts.Count > 0)
+            else
             {
-                foreach (var post in user.Posts)
-                {
-                    // decrementam nr de mesaje din forum si subforum
-                    Subforum sf = db.Subforums.Find(post.SubforumId);
-                    Forum f = db.Forums.Find(sf.ForumId);
-                    sf.MsgCount--;
-                    f.MsgCount--;
-                    db.Posts.Remove(post);
-                }
+                return Redirect("/Identity/Account/Login");
             }
-
-            db.ApplicationUsers.Remove(user);
-            db.SaveChanges();
-
-            return RedirectToAction("Index");
         }
 
         [NonAction]
@@ -159,7 +203,7 @@ namespace ForumApp.Controllers
             var roles = from role in db.Roles
                         select role;
 
-            foreach(var role in roles)
+            foreach (var role in roles)
             {
                 selectList.Add(new SelectListItem
                 {
